pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-mcp"
version = "0.1.0"
description = "Intelligent middleware for MCP servers — observe, reason, evolve"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.9",
    "aiohttp>=3.9",
    "aiosqlite>=0.19",
    "cyclopts>=4.0",
    "python-dotenv>=1.0",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
